TaxSEA: Taxon Set Enrichment Analysis

Modular TaxSEA implementation supporting enrichment (KS) and ORA (Fisher). Provide either taxon_ranks for enrichment or input_taxa for ORA.

Usage

TaxSEA(
  taxon_ranks = NULL,
  input_taxa = NULL,
  mode = NULL,
  lookup_missing = FALSE,
  min_set_size = 5,
  max_set_size = 300,
  custom_db = NULL,
  bugsigdb = FALSE
)

Arguments

taxon_ranks

Named numeric vector of statistics (e.g., log2 fold changes). Required for enrichment.

input_taxa

Character vector of taxa to treat as "hits"/selected taxa. Required for ORA.

mode

Character. One of "enrichment" or "ora". If NULL, inferred from which input is provided.

lookup_missing

Logical indicating whether to fetch missing NCBI IDs. Default is FALSE.

min_set_size

Minimum size of taxon sets to include in the analysis. Default is 5.

max_set_size

Maximum size of taxon sets to include in the analysis. Default is 300.

custom_db

A user-provided list of taxon sets. If NULL (default), the built-in database is used.

bugsigdb

Logical; whether to augment the built-in database with BugSigDB signatures, which are downloaded (and cached) at run time via bugsigdbr. Default is FALSE. Each set is tested against all the other taxa covered by the sets being analysed, and BugSigDB adds many extra taxa, so including it changes the p-values of every set, not only the BugSigDB ones. BugSigDB is also updated independently of TaxSEA. Set to TRUE to include it, and record the BugSigDB version used. If the download fails, a warning is given and the analysis continues without BugSigDB. Ignored when custom_db is supplied.

Value

A list of data frames with taxon set results.

Examples

data("TaxSEA_test_data")
res <- TaxSEA(taxon_ranks = TaxSEA_test_data)
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
head(res$All_databases)
#>                                        taxonSetName median_rank_of_set_members
#> 1             GMRepoV2_Decreased_in_Liver_Cirrhosis                    -3.6435
#> 2                                     mBodyMap_Oral                     1.7750
#> 3                    BacDive_Uses_nitrate_reduction                     1.7750
#> 4 GMRepoV2_Decreased_in_Inflammatory_Bowel_Diseases                    -3.3870
#> 5                           BloSSUM_taxa_Carter2023                    -2.8795
#> 6 GMRepoV2_Increased_in_Inflammatory_Bowel_Diseases                     1.4650
#>         PValue Test_statistic         FDR
#> 1 4.159210e-06      0.4534706 0.001301833
#> 2 1.568875e-05      0.7028986 0.002455290
#> 3 5.374713e-05      0.6259755 0.005607617
#> 4 9.592476e-05      0.4666667 0.007506113
#> 5 3.163143e-04      0.3878943 0.015138713
#> 6 3.607350e-04      0.4107946 0.015138713
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 TaxonSet
#> 1 Faecalibacterium_prausnitzii, Bacteroides_uniformis, Roseburia_hominis, Alistipes_putredinis, Bacteroides_dorei, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Bacteroides_faecis, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Bacteroides_xylanisolvens, Collinsella_aerofaciens, Bacteroides_massiliensis, Odoribacter_splanchnicus, Eubacterium_hallii, Bilophila_wadsworthia, Eubacterium_ramulus, Eubacterium_ventriosum, Coprococcus_catus, Alistipes_shahii, Roseburia_intestinalis, Eubacterium_siraeum, Intestinibacter_bartlettii, Ruminococcus_bromii, Bifidobacterium_adolescentis, Coprococcus_comes, Bacteroides_cellulosilyticus, Ruminococcus_callidus, Coprococcus_eutactus, Ruminococcus_lactaris, Bacteroides_finegoldii, Barnesiella_intestinihominis, Alistipes_indistinctus, Alistipes_onderdonkii, Bacteroides_intestinalis
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Streptococcus_salivarius, Haemophilus_parainfluenzae, Streptococcus_parasanguinis, Actinomyces_oris, Veillonella_parvula, Veillonella_dispar, Streptococcus_mitis, Streptococcus_sanguinis, Rothia_mucilaginosa, Actinomyces_sp_HPA0247, Actinomyces_sp_ICM47
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Escherichia_coli, Haemophilus_parainfluenzae, Clostridium_lavalense, Hungatella_hathewayi, Clostridium_bolteae, Actinomyces_oris, Veillonella_parvula, Veillonella_dispar, Veillonella_infantium, Veillonella_atypica, Rothia_mucilaginosa, Klebsiella_pneumoniae, Clostridium_saccharolyticum
#> 4                                                                                                                                                                                                                                                                                                                                               Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Akkermansia_muciniphila, Bacteroides_ovatus, Bacteroides_faecis, Dorea_formicigenerans, Roseburia_inulinivorans, Bacteroides_caccae, Collinsella_aerofaciens, Alistipes_finegoldii, Bacteroides_massiliensis, Odoribacter_splanchnicus, Eubacterium_ramulus, Adlercreutzia_equolifaciens, Coprococcus_catus, Clostridium_leptum, Parabacteroides_merdae, Ruminococcus_bromii, Ruminococcus_torques, Dialister_invisus, Bacteroides_fragilis, Alistipes_onderdonkii, Bacteroides_intestinalis
#> 5                                    Bacteroides_uniformis, Alistipes_putredinis, Bacteroides_dorei, Eubacterium_rectale, Anaerostipes_hadrus, Akkermansia_muciniphila, Bacteroides_ovatus, Bacteroides_faecis, Parasutterella_excrementihominis, Bacteroides_xylanisolvens, Bacteroides_caccae, Parabacteroides_distasonis, Alistipes_finegoldii, Bacteroides_massiliensis, Odoribacter_splanchnicus, Bacteroides_thetaiotaomicron, Bilophila_wadsworthia, Phascolarctobacterium_faecium, Lawsonibacter_asaccharolyticus, Alistipes_shahii, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Streptococcus_thermophilus, Parabacteroides_merdae, Bacteroides_stercoris, Bifidobacterium_adolescentis, Bifidobacterium_bifidum, Ruminococcus_torques, Bacteroides_cellulosilyticus, Ruminococcus_lactaris, Bacteroides_finegoldii, Barnesiella_intestinihominis, Alistipes_indistinctus, Alistipes_onderdonkii
#> 6                                                                                                                                                                                                           Anaerostipes_hadrus, Bacteroides_thetaiotaomicron, Streptococcus_salivarius, Bilophila_wadsworthia, Escherichia_coli, Gordonibacter_pamelaeae, Flavonifractor_plautii, Streptococcus_parasanguinis, Streptococcus_thermophilus, Eggerthella_lenta, Hungatella_hathewayi, Blautia_producta, Bifidobacterium_bifidum, Lactococcus_lactis, Ruminococcus_callidus, Veillonella_parvula, Clostridium_spiroforme, Bifidobacterium_dentium, Clostridium_scindens, Streptococcus_mitis, Rothia_mucilaginosa, Butyricicoccus_pullicaecorum, Bifidobacterium_animalis, Bifidobacterium_breve, Tyzzerella_nexilis, Ruminococcaceae_bacterium_D16, Enterococcus_faecium, Streptococcus_vestibularis, Anaerostipes_caccae

# \donttest{
# Include BugSigDB signatures (downloaded at run time)
res_bsdb <- TaxSEA(taxon_ranks = TaxSEA_test_data, bugsigdb = TRUE)
#> 
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
head(res_bsdb$BugSigDB)
#>   PubMedID
#> 1 39237540
#> 2 39237540
#> 3 39237540
#> 4 39237540
#> 5 35692191
#> 6 38968070
#>                                                                                                                                                                                 BugSigDB_ID
#> 1                                                                                 bsdb_bsdb:39237540/12/2_Aging:All-cohorts-(Increased-age-group)_vs_All-cohorts-(Decreased-age-group)_DOWN
#> 2                                                                                 bsdb_bsdb:39237540/13/2_Aging:All-cohorts-(Increased-age-group)_vs_All-cohorts-(Decreased-age-group)_DOWN
#> 3                                                                                bsdb_bsdb:39237540/15/2_Aging:All-cohorts-(Increased-age-group)_vs_All-Cohorts-(Decreasing-age-group)_DOWN
#> 4                                                                                  bsdb_bsdb:39237540/4/2_Aging:All-Cohorts-(Increased-age-group)_vs_All-Cohorts-(Decreased-age-group)_DOWN
#> 5                                                                                         bsdb_bsdb:35692191/2/2_Crohn's-disease:Crohn's-disease-patient_vs_Healthy-control-individual_DOWN
#> 6 bsdb_bsdb:38968070/11/1_Socioeconomic-status:Subjects-from-the-wealthier-households-based-on-economic-factors_vs_Subjects-from-the-least-well-off-households-based-on-economic-factors_UP
#>   median_rank_of_set_members       PValue Test_statistic          FDR
#> 1                     -3.753 6.403851e-06      0.5809088 0.0006774738
#> 2                     -3.753 6.403851e-06      0.5809088 0.0006774738
#> 3                     -3.753 6.403851e-06      0.5809088 0.0006774738
#> 4                     -3.824 7.742558e-06      0.5426313 0.0006774738
#> 5                     -4.711 4.351745e-05      0.7639752 0.0030462214
#> 6                     -2.893 6.938443e-05      0.4099379 0.0037779329
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    TaxonSet
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                       Faecalibacterium_prausnitzii, Roseburia_hominis, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Eubacterium_hallii, Eubacterium_ventriosum, Coprococcus_catus, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Ruminococcus_bromii, Coprococcus_comes, Ruminococcus_torques, Ruminococcus_callidus, Ruminococcus_lactaris
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                       Faecalibacterium_prausnitzii, Roseburia_hominis, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Eubacterium_hallii, Eubacterium_ventriosum, Coprococcus_catus, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Ruminococcus_bromii, Coprococcus_comes, Ruminococcus_torques, Ruminococcus_callidus, Ruminococcus_lactaris
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                       Faecalibacterium_prausnitzii, Roseburia_hominis, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Eubacterium_hallii, Eubacterium_ventriosum, Coprococcus_catus, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Ruminococcus_bromii, Coprococcus_comes, Ruminococcus_torques, Ruminococcus_callidus, Ruminococcus_lactaris
#> 4                                                                                                                                                                                                                                                                                                                                           Faecalibacterium_prausnitzii, Roseburia_hominis, Eubacterium_rectale, Ruminococcus_gnavus, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Collinsella_aerofaciens, Eubacterium_hallii, Eubacterium_ventriosum, Coprococcus_catus, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Ruminococcus_bromii, Bifidobacterium_adolescentis, Coprococcus_comes, Ruminococcus_torques, Ruminococcus_callidus, Ruminococcus_lactaris
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Faecalibacterium_prausnitzii, Alistipes_putredinis, Dorea_longicatena, Roseburia_inulinivorans, Collinsella_aerofaciens, Eubacterium_siraeum, Parabacteroides_merdae, Ruminococcus_bromii
#> 6 Bacteroides_uniformis, Alistipes_putredinis, Bacteroides_dorei, Fusicatenibacter_saccharivorans, Blautia_wexlerae, Akkermansia_muciniphila, Bacteroides_ovatus, Dorea_formicigenerans, Roseburia_inulinivorans, Bacteroides_xylanisolvens, Bacteroides_caccae, Parabacteroides_distasonis, Collinsella_aerofaciens, Alistipes_finegoldii, Bacteroides_massiliensis, Odoribacter_splanchnicus, Bacteroides_thetaiotaomicron, Bilophila_wadsworthia, Phascolarctobacterium_faecium, Flavonifractor_plautii, Alistipes_shahii, Bacteroides_vulgatus, Bifidobacterium_longum, Parabacteroides_merdae, Bacteroides_stercoris, Ruminococcus_torques, Bacteroides_cellulosilyticus, Lactococcus_lactis, Coprococcus_eutactus, Ruminococcus_lactaris, Bacteroides_fragilis, Bacteroides_finegoldii, Barnesiella_intestinihominis, Alistipes_onderdonkii, Bacteroides_intestinalis
#>   PubMedID.1
#> 1   39237540
#> 2   39237540
#> 3   39237540
#> 4   39237540
#> 5   35692191
#> 6   38968070
# }

# ORA example (toy): treat taxa with positive values as "hits"
hits <- names(TaxSEA_test_data)
res_ora <- TaxSEA(input_taxa = hits, mode = "ora")
head(res_ora$All_databases)
#>                                                       taxonSetName       PValue
#> 1                  Valles-Colomer2019_MGB006.Glutamate.synthesis.I 7.442269e-62
#> 2                 Valles-Colomer2019_MGB007.Glutamate.synthesis.II 1.065220e-61
#> 3 Valles-Colomer2019_MGB029.ClpB.(ATP-dependent.chaperone.protein) 1.383936e-60
#> 4   Valles-Colomer2019_MGB036.S-Adenosylmethionine.(SAM).synthesis 3.382557e-60
#> 5                    Valles-Colomer2019_MGB043.Acetate.synthesis.I 2.852085e-59
#> 6            Valles-Colomer2019_MGB033.Quinolinic.acid.degradation 1.052201e-57
#>   Odds_ratio          FDR
#> 1   45.11695 7.046430e-59
#> 2   44.78065 7.046430e-59
#> 3   38.35351 6.103159e-58
#> 4   37.69384 1.118781e-57
#> 5   38.52611 7.546618e-57
#> 6   35.83523 2.320103e-55
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               TaxonSet
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Gordonibacter_pamelaeae, Clostridium_citroniae, Dialister_succinatiphilus, Paraprevotella_clara, Lactobacillus_ruminis, Clostridium_leptum, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Bacteroides_ovatus, Megamonas_funiformis, Methanobrevibacter_smithii, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lacticaseibacillus_rhamnosus, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Escherichia_sp., Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_brevis, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Corynebacterium_ammoniagenes, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Enterobacter_cancerogenus, 1872444, Paenibacillus_sp., NA, NA, Burkholderia oklahomensis, Providencia_rustigianii, 54061, NA, 1872530, Bifidobacterium_sp., NA, 205, 1896336, Clostridium_clostridioforme, Clostridium_methylpentosum, Coprobacillus_sp., 2049024, Desulfovibrio_sp., Dysgonomonas_mossii, Helicobacter bizzozeronii, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, 576, Methanosphaera_stadtmanae, Providencia_alcalifaciens, 625, Streptococcus_sp., Victivallis_vadensis, Anaerofustis_stercorihominis, NA, Eubacterium_sp., Fusobacterium_necrophorum, 1977903, Weissella_paramesenteroides
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Parasutterella_excrementihominis, Clostridium_leptum, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_fermentum, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Ruminococcus_gnavus, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lacticaseibacillus_rhamnosus, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Acinetobacter_baumannii, Bacillus_cereus, Enterococcus_saccharolyticus, Campylobacter_jejuni, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, 1872444, Bacillus smithii, Paenibacillus_sp., NA, NA, Vibrio_cholerae, 678, Burkholderia oklahomensis, Providencia_rustigianii, 54061, NA, 1872530, Bifidobacterium_sp., NA, Campylobacter_coli, Campylobacter_upsaliensis, 1896336, Clostridium_clostridioforme, Clostridium_methylpentosum, Coprobacillus_sp., 2049024, Desulfovibrio_sp., Dysgonomonas_mossii, 576, NA, Providencia_alcalifaciens, 625, Victivallis_vadensis, Anaerofustis_stercorihominis, NA, Eubacterium_sp., 1977903, Fusobacterium_gonidiaformans, NA
#> 3                                                            Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Gordonibacter_pamelaeae, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Lactobacillus_crispatus, Parasutterella_excrementihominis, Clostridium_leptum, Limosilactobacillus_reuteri, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_delbrueckii, Lactobacillus_johnsonii, Lactobacillus_fermentum, Lactobacillus_gasseri, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lactobacillus_acidophilus, Lacticaseibacillus_rhamnosus, Pediococcus_acidilactici, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Sutterella_parvirubra, Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Finegoldia_magna, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Bacillus_cereus, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_jejuni, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Lactobacillus_helveticus, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Lactobacillus_antri, Leuconostoc_mesenteroides, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Corynebacterium_ammoniagenes, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Lactobacillus_amylolyticus, Lactobacillus_ultunensis, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, 1872444, Bacillus smithii, Paenibacillus_sp., NA, NA, Vibrio_cholerae, 678, Burkholderia oklahomensis, Providencia_rustigianii, 54061, NA, NA, 1872530, Bifidobacterium_sp., NA, Campylobacter_coli, 205, Campylobacter_upsaliensis, 1896336, Clostridium_clostridioforme, Clostridium_methylpentosum, Coprobacillus_sp., 2049024, Desulfovibrio_sp., Dysgonomonas_mossii, Helicobacter bizzozeronii, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, 576, NA, Providencia_alcalifaciens, 625, Streptococcus_sp., Victivallis_vadensis, Anaerofustis_stercorihominis, Eubacterium_sp., Fusobacterium_necrophorum, 1977903, Weissella_paramesenteroides, Fusobacterium_gonidiaformans, NA, Lactobacillus_fructivorans, Parvimonas_micra, Bartonella_quintana, 1591
#> 4 Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Gordonibacter_pamelaeae, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Lactobacillus_crispatus, Parasutterella_excrementihominis, Clostridium_leptum, Limosilactobacillus_reuteri, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Methanobrevibacter_smithii, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_delbrueckii, Lactobacillus_johnsonii, Lactobacillus_fermentum, Lactobacillus_gasseri, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lactobacillus_acidophilus, Lacticaseibacillus_rhamnosus, Pediococcus_acidilactici, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Sutterella_parvirubra, Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Finegoldia_magna, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Bacillus_cereus, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_jejuni, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Lactobacillus_helveticus, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Lactobacillus_antri, Leuconostoc_mesenteroides, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Corynebacterium_ammoniagenes, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Lactobacillus_amylolyticus, Lactobacillus_ultunensis, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, 1872444, Bacillus smithii, Paenibacillus_sp., NA, NA, Vibrio_cholerae, 678, Burkholderia oklahomensis, Providencia_rustigianii, 54061, NA, NA, 1872530, Bifidobacterium_sp., NA, Campylobacter_coli, 205, Campylobacter_upsaliensis, 1896336, Clostridium_clostridioforme, Clostridium_methylpentosum, Coprobacillus_sp., 2049024, Desulfovibrio_sp., Dysgonomonas_mossii, Helicobacter bizzozeronii, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, 576, Methanosphaera_stadtmanae, NA, Providencia_alcalifaciens, 625, Streptococcus_sp., Victivallis_vadensis, Anaerofustis_stercorihominis, NA, Eubacterium_sp., Fusobacterium_necrophorum, 1977903, Weissella_paramesenteroides, Fusobacterium_gonidiaformans, NA, Lactobacillus_fructivorans, Parvimonas_micra, Bartonella_quintana, 1591
#> 5                                                                                                                                                                                                                                                                                                                                                                                                             Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Clostridium_citroniae, Paraprevotella_clara, Lactobacillus_ruminis, Lactobacillus_crispatus, Clostridium_leptum, Limosilactobacillus_reuteri, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_delbrueckii, Lactobacillus_johnsonii, Lactobacillus_fermentum, Lactobacillus_gasseri, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lactobacillus_acidophilus, Lacticaseibacillus_rhamnosus, Pediococcus_acidilactici, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Finegoldia_magna, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Bacillus_cereus, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_jejuni, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Lactobacillus_helveticus, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Lactobacillus_antri, Leuconostoc_mesenteroides, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Corynebacterium_ammoniagenes, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Lactobacillus_amylolyticus, Lactobacillus_ultunensis, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Enterobacter_cancerogenus, 1872444, Paenibacillus_sp., NA, NA, Vibrio_cholerae, 678, Burkholderia oklahomensis, Providencia_rustigianii, 54061, NA, NA, 1872530, Bifidobacterium_sp., NA, Campylobacter_coli, 205, Campylobacter_upsaliensis, 1896336, Clostridium_clostridioforme, Clostridium_methylpentosum, Coprobacillus_sp., 2049024, Desulfovibrio_sp., Dysgonomonas_mossii, Helicobacter bizzozeronii, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, 576, NA, Providencia_alcalifaciens, 625, Streptococcus_sp., Victivallis_vadensis, Anaerofustis_stercorihominis, NA, Eubacterium_sp., Fusobacterium_necrophorum, Weissella_paramesenteroides, Fusobacterium_gonidiaformans, NA, Lactobacillus_fructivorans, Parvimonas_micra, 1591
#> 6                                                                                                           Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Lactobacillus_crispatus, Parasutterella_excrementihominis, Clostridium_leptum, Limosilactobacillus_reuteri, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Methanobrevibacter_smithii, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_delbrueckii, Lactobacillus_johnsonii, Lactobacillus_fermentum, Lactobacillus_gasseri, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lactobacillus_acidophilus, Lacticaseibacillus_rhamnosus, Pediococcus_acidilactici, Pseudomonas_sp., Sutterella_parvirubra, Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Finegoldia_magna, Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Bacillus_cereus, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_jejuni, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Lactobacillus_helveticus, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Lactobacillus_antri, Leuconostoc_mesenteroides, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Corynebacterium_ammoniagenes, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Lactobacillus_amylolyticus, Lactobacillus_ultunensis, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, 1872444, Bacillus smithii, Paenibacillus_sp., NA, NA, Vibrio_cholerae, 678, Burkholderia oklahomensis, Providencia_rustigianii, 54061, NA, NA, 1872530, Bifidobacterium_sp., NA, Campylobacter_coli, 205, Campylobacter_upsaliensis, 1896336, Clostridium_clostridioforme, Clostridium_methylpentosum, Coprobacillus_sp., 2049024, Desulfovibrio_sp., Dysgonomonas_mossii, Helicobacter bizzozeronii, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, 576, Methanosphaera_stadtmanae, NA, Providencia_alcalifaciens, 625, Streptococcus_sp., Victivallis_vadensis, Anaerofustis_stercorihominis, Eubacterium_sp., Fusobacterium_necrophorum, 1977903, Weissella_paramesenteroides, Fusobacterium_gonidiaformans, NA, Lactobacillus_fructivorans, Parvimonas_micra, Bartonella_quintana, 1591