TaxSEA: Taxon Set Enrichment Analysis

TaxSEA is an R package for Taxon Set Enrichment Analysis, which utilises a Kolmogorov-Smirnov test analyses to investigate differential abundance analysis output for whether there are alternations in a-priori defined sets of taxa from public databases (BugSigDB, MiMeDB, GutMGene, mBodyMap, BacDive and GMRepoV2) and collated from the literature. TaxSEA takes as input a list of taxonomic identifiers (e.g. species names, NCBI IDs etc.) and a rank (E.g. fold change, correlation coefficient). TaxSEA be applied to any microbiota taxonomic profiling technology (array-based, 16S rRNA gene sequencing, shotgun metagenomics & metatranscriptomics etc.) and enables researchers to rapidly contextualize their findings within the broader literature to accelerate interpretation of results.

See also

Author

Maintainer: Feargal Ryan feargalr@gmail.com (ORCID) (funding: Supported by NHMRC Investigator Grant) [funder]

Authors:

Feargal Ryan feargalr@gmail.com (ORCID) (funding: Supported by NHMRC Investigator Grant) [funder]