TaxSEA: Taxon Set Enrichment Analysis

Modular TaxSEA implementation supporting enrichment (KS) and ORA (Fisher). Provide either taxon_ranks for enrichment or input_taxa for ORA.

Usage

TaxSEA(
  taxon_ranks = NULL,
  input_taxa = NULL,
  mode = NULL,
  lookup_missing = FALSE,
  min_set_size = 5,
  max_set_size = 300,
  custom_db = NULL
)

Arguments

taxon_ranks

Named numeric vector of statistics (e.g., log2 fold changes). Required for enrichment.

input_taxa

Character vector of taxa to treat as "hits"/selected taxa. Required for ORA.

mode

Character. One of "enrichment" or "ora". If NULL, inferred from which input is provided.

lookup_missing

Logical indicating whether to fetch missing NCBI IDs. Default is FALSE.

min_set_size

Minimum size of taxon sets to include in the analysis. Default is 5.

max_set_size

Maximum size of taxon sets to include in the analysis. Default is 100.

custom_db

A user-provided list of taxon sets. If NULL (default), the built-in database is used.

Value

A list of data frames with taxon set results.

Examples

data("TaxSEA_test_data")
res <- TaxSEA(taxon_ranks = TaxSEA_test_data)
#> 
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
#> Warning: p-value will be approximate in the presence of ties
head(res$All_databases)
#>                                                                                                        taxonSetName
#> 1       bsdb_bsdb:562/3/2_Health-study-participation:feces-from-healthy-adult_vs_none---50%-prevalence-threshold_NA
#> 2       bsdb_bsdb:562/9/2_Health-study-participation:feces-from-healthy-adult_vs_none---70%-prevalence-threshold_NA
#> 3                                                                             GMRepoV2_Decreased_in_Liver_Cirrhosis
#> 4 bsdb_bsdb:562/5/2_Health-study-participation:oral-cavity-from-healthy-adult_vs_none---50%-prevalence-threshold_NA
#> 5         bsdb_bsdb:39237540/12/2_Aging:All-cohorts-(Increased-age-group)_vs_All-cohorts-(Decreased-age-group)_DOWN
#> 6         bsdb_bsdb:39237540/13/2_Aging:All-cohorts-(Increased-age-group)_vs_All-cohorts-(Decreased-age-group)_DOWN
#>   median_rank_of_set_members       PValue Test_statistic          FDR
#> 1                    -3.3370 1.155649e-07      0.4199597 3.951813e-05
#> 2                    -3.8735 1.393938e-07      0.5104167 3.951813e-05
#> 3                    -3.6435 2.699126e-06      0.4532895 5.101348e-04
#> 4                     1.8330 6.409955e-06      0.6395833 5.677407e-04
#> 5                    -3.7530 7.009145e-06      0.5786184 5.677407e-04
#> 6                    -3.7530 7.009145e-06      0.5786184 5.677407e-04
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            TaxonSet
#> 1 Faecalibacterium_prausnitzii, Bacteroides_uniformis, Roseburia_hominis, Alistipes_putredinis, Bacteroides_dorei, Eubacterium_rectale, Fusicatenibacter_saccharivorans, Dorea_longicatena, Agathobaculum_butyriciproducens, Blautia_obeum, Anaerostipes_hadrus, Ruminococcus_bicirculans, Blautia_wexlerae, Eubacterium_eligens, Akkermansia_muciniphila, Bacteroides_ovatus, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Bacteroides_xylanisolvens, Bacteroides_caccae, Parabacteroides_distasonis, Collinsella_aerofaciens, Alistipes_finegoldii, Gemmiger_formicilis, Odoribacter_splanchnicus, Eubacterium_hallii, Bacteroides_thetaiotaomicron, Streptococcus_salivarius, Ruthenibacterium_lactatiformans, Bilophila_wadsworthia, Escherichia_coli, Eubacterium_ramulus, Adlercreutzia_equolifaciens, Eubacterium_ventriosum, Coprococcus_catus, Gordonibacter_pamelaeae, Flavonifractor_plautii, Alistipes_shahii, Streptococcus_parasanguinis, Bacteroides_vulgatus, Oscillibacter_sp_57_20, Roseburia_faecis, Roseburia_intestinalis, Eubacterium_siraeum, Lachnospira_pectinoschiza, Eubacterium_sp_CAG_38, Clostridium_leptum, Intestinibacter_bartlettii, Prevotella_copri, Bifidobacterium_longum, Parabacteroides_merdae, Oscillibacter_sp_CAG_241, Ruminococcus_bromii, Collinsella_intestinalis, Bifidobacterium_adolescentis, Coprococcus_comes, Ruminococcus_torques, Roseburia_sp_CAG_471, Barnesiella_intestinihominis, Firmicutes_bacterium_CAG_83, Butyricimonas_virosa
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Faecalibacterium_prausnitzii, Bacteroides_uniformis, Roseburia_hominis, Alistipes_putredinis, Eubacterium_rectale, Fusicatenibacter_saccharivorans, Dorea_longicatena, Agathobaculum_butyriciproducens, Blautia_obeum, Anaerostipes_hadrus, Blautia_wexlerae, Eubacterium_eligens, Bacteroides_ovatus, Dorea_formicigenerans, Roseburia_inulinivorans, Parabacteroides_distasonis, Collinsella_aerofaciens, Alistipes_finegoldii, Gemmiger_formicilis, Odoribacter_splanchnicus, Eubacterium_hallii, Bacteroides_thetaiotaomicron, Streptococcus_salivarius, Ruthenibacterium_lactatiformans, Eubacterium_ramulus, Coprococcus_catus, Flavonifractor_plautii, Alistipes_shahii, Bacteroides_vulgatus, Oscillibacter_sp_57_20, Roseburia_faecis, Roseburia_intestinalis, Bifidobacterium_longum, Coprococcus_comes, Ruminococcus_torques, Firmicutes_bacterium_CAG_83
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Faecalibacterium_prausnitzii, Bacteroides_uniformis, Roseburia_hominis, Alistipes_putredinis, Bacteroides_dorei, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Bacteroides_faecis, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Bacteroides_xylanisolvens, Collinsella_aerofaciens, Bacteroides_massiliensis, Odoribacter_splanchnicus, Eubacterium_hallii, Bilophila_wadsworthia, Eubacterium_ramulus, Eubacterium_ventriosum, Coprococcus_catus, Alistipes_shahii, Roseburia_intestinalis, Eubacterium_siraeum, Intestinibacter_bartlettii, Ruminococcus_bromii, Bifidobacterium_adolescentis, Coprococcus_comes, Bacteroides_cellulosilyticus, Ruminococcus_callidus, Coprococcus_eutactus, Ruminococcus_lactaris, Bacteroides_finegoldii, Barnesiella_intestinihominis, Alistipes_indistinctus, Alistipes_onderdonkii, Bacteroides_intestinalis
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Streptococcus_salivarius, Haemophilus_parainfluenzae, Streptococcus_parasanguinis, Streptococcus_australis, Actinomyces_oris, Veillonella_parvula, Veillonella_dispar, Veillonella_infantium, Veillonella_atypica, Streptococcus_mitis, Streptococcus_sanguinis, Rothia_mucilaginosa, Veillonella_sp_T11011_6, Actinomyces_sp_HMSC035G02, Actinomyces_sp_ICM47
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Faecalibacterium_prausnitzii, Roseburia_hominis, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Eubacterium_hallii, Eubacterium_ventriosum, Coprococcus_catus, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Ruminococcus_bromii, Coprococcus_comes, Ruminococcus_torques, Ruminococcus_callidus, Ruminococcus_lactaris
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Faecalibacterium_prausnitzii, Roseburia_hominis, Eubacterium_rectale, Dorea_longicatena, Blautia_obeum, Eubacterium_eligens, Dorea_formicigenerans, Roseburia_inulinivorans, Parasutterella_excrementihominis, Eubacterium_hallii, Eubacterium_ventriosum, Coprococcus_catus, Roseburia_intestinalis, Bifidobacterium_pseudocatenulatum, Ruminococcus_bromii, Coprococcus_comes, Ruminococcus_torques, Ruminococcus_callidus, Ruminococcus_lactaris

# ORA example (toy): treat taxa with positive values as "hits"
hits <- names(TaxSEA_test_data)
res_ora <- TaxSEA(input_taxa = hits, mode = "ora")
#> Using cached version from 2026-01-14 23:25:30
head(res_ora$All_databases)
#>                                                                                                  taxonSetName
#> 1 bsdb_bsdb:562/3/2_Health-study-participation:feces-from-healthy-adult_vs_none---50%-prevalence-threshold_NA
#> 2                                                             Valles-Colomer2019_MGB006.Glutamate.synthesis.I
#> 3                                                            Valles-Colomer2019_MGB007.Glutamate.synthesis.II
#> 4                                   bsdb_bsdb:37438797/1/2_COVID-19:Severe-COVID-19_vs_Moderate-COVID-19_DOWN
#> 5                                            Valles-Colomer2019_MGB029.ClpB.(ATP-dependent.chaperone.protein)
#> 6                                              Valles-Colomer2019_MGB036.S-Adenosylmethionine.(SAM).synthesis
#>         PValue Odds_ratio          FDR
#> 1 5.370863e-84  370.73342 2.304100e-80
#> 2 1.357889e-52   27.54324 2.762292e-49
#> 3 1.931673e-52   27.34134 2.762292e-49
#> 4 4.155598e-51  164.68510 4.456879e-48
#> 5 8.685139e-51   22.96775 7.451849e-48
#> 6 2.083910e-50   22.57092 1.489996e-47
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        TaxonSet
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Streptococcus_salivarius, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Coprococcus_catus, Eubacterium_ramulus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Gordonibacter_pamelaeae, Parasutterella_excrementihominis, Streptococcus_parasanguinis, Clostridium_leptum, Bacteroides_dorei, Bacteroides_caccae, Parabacteroides_merdae, Escherichia_coli, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_uniformis, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Roseburia_hominis, Alistipes_shahii, Bacteroides_xylanisolvens, Anaerostipes_hadrus, Bilophila_wadsworthia, Ruminococcus_bromii, Barnesiella_intestinihominis, Adlercreutzia_equolifaciens, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Bacteroides_ovatus, Eubacterium_siraeum, Akkermansia_muciniphila, Alistipes_finegoldii, Odoribacter_splanchnicus, Bacteroides_distasonis, Clostridium_orbiscindens, Butyricimonas_virosa , Fusicatenibacter_saccharivorans, Blautia_wexlerae, Ruthenibacterium_lactatiformans, Roseburia_faecis, Agathobaculum_butyriciproducens, Gemmiger_formicilis, Firmicutes_bacterium_CAG_83, Oscillibacter_sp_57_20, Oscillibacter_sp_CAG_241, Ruminococcus_bicirculans, NA, Eubacterium_sp_CAG_38, Lachnospira_pectinoschiza, Roseburia_sp_CAG_471, Firmicutes_bacterium_CAG_110, NA
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Gordonibacter_pamelaeae, Clostridium_citroniae, Dialister_succinatiphilus, Paraprevotella_clara, Lactobacillus_ruminis, Clostridium_leptum, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Bacteroides_ovatus, Megamonas_funiformis, Methanobrevibacter_smithii, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lacticaseibacillus_rhamnosus, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Escherichia_sp., Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_brevis, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Corynebacterium_ammoniagenes, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Enterobacter_cancerogenus, Providencia_alcalifaciens, Victivallis_vadensis, Anaerofustis_stercorihominis, Dysgonomonas_mossii, Fusobacterium_necrophorum, NA, Paenibacillus_sp., NA, NA, NA, Providencia_rustigianii, NA, NA, NA, Bifidobacterium_sp., NA, NA, NA, NA, Clostridium_methylpentosum, Coprobacillus_sp., NA, Desulfovibrio_sp., NA, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, NA, Methanosphaera_stadtmanae, NA, Streptococcus_sp., NA, Eubacterium_sp., NA, Weissella_paramesenteroides
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Parasutterella_excrementihominis, Clostridium_leptum, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_fermentum, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Ruminococcus_gnavus, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lacticaseibacillus_rhamnosus, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Acinetobacter_baumannii, Bacillus_cereus, Enterococcus_saccharolyticus, Campylobacter_jejuni, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, Providencia_alcalifaciens, Victivallis_vadensis, Anaerofustis_stercorihominis, Dysgonomonas_mossii, Campylobacter_upsaliensis, Fusobacterium_gonidiaformans, NA, NA, Paenibacillus_sp., NA, NA, Vibrio_cholerae, NA, NA, Providencia_rustigianii, NA, NA, NA, Bifidobacterium_sp., NA, Campylobacter_coli, NA, NA, Clostridium_methylpentosum, Coprobacillus_sp., NA, Desulfovibrio_sp., NA, NA, NA, NA, Eubacterium_sp., NA, NA
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Streptococcus_salivarius, Streptococcus_thermophilus, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Coprococcus_catus, Eubacterium_ramulus, Eubacterium_ventriosum, Lactococcus_lactis, Streptococcus_parasanguinis, Clostridium_leptum, Roseburia_inulinivorans, Roseburia_intestinalis, Roseburia_hominis, Bifidobacterium_pseudocatenulatum, Oscillibacter_, Anaerostipes_hadrus, Ruminococcus_bromii, Adlercreutzia_equolifaciens, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Eubacterium_siraeum, Rothia_mucilaginosa, Streptococcus_gordonii, Gemella_sanguinis, Streptococcus_mitis, Collinsella_stercoris, Asaccharobacter_celatus, Streptococcus_oralis, Actinomyces_sp._HPA0247, Actinomyces_oris, Schaalia_odontolytica, Monoglobus_pectinilyticus, Fusicatenibacter_saccharivorans, Blautia_wexlerae, Roseburia_faecis, Agathobaculum_butyriciproducens, Gemmiger_formicilis, Ruminococcus_bicirculans, Firmicutes_bacterium_CAG_95, Firmicutes_bacterium_CAG_94
#> 5                                                            Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Gordonibacter_pamelaeae, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Lactobacillus_crispatus, Parasutterella_excrementihominis, Clostridium_leptum, Limosilactobacillus_reuteri, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_delbrueckii, Lactobacillus_johnsonii, Lactobacillus_fermentum, Lactobacillus_gasseri, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lactobacillus_acidophilus, Lacticaseibacillus_rhamnosus, Pediococcus_acidilactici, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Sutterella_parvirubra, Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Finegoldia_magna, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Bacillus_cereus, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_jejuni, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Lactobacillus_helveticus, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Lactobacillus_antri, Leuconostoc_mesenteroides, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Corynebacterium_ammoniagenes, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Lactobacillus_amylolyticus, Lactobacillus_ultunensis, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, Providencia_alcalifaciens, Victivallis_vadensis, Parvimonas_micra, Anaerofustis_stercorihominis, Dysgonomonas_mossii, Lactobacillus_fructivorans, Campylobacter_upsaliensis, Fusobacterium_necrophorum, Fusobacterium_gonidiaformans, NA, NA, Paenibacillus_sp., NA, NA, Vibrio_cholerae, NA, NA, Providencia_rustigianii, NA, NA, NA, NA, Bifidobacterium_sp., NA, Campylobacter_coli, NA, NA, NA, Clostridium_methylpentosum, Coprobacillus_sp., NA, Desulfovibrio_sp., NA, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, NA, NA, NA, Streptococcus_sp., Eubacterium_sp., NA, Weissella_paramesenteroides, NA, Bartonella_quintana, NA
#> 6 Eubacterium_rectale, Bifidobacterium_stercoris, Bifidobacterium_longum, Eubacterium_eligens, Anaerobutyricum_hallii, Ruminococcus_obeum, Dorea_longicatena, Ruminococcus_torques, Coprococcus_comes, Dorea_formicigenerans, Bifidobacterium_catenulatum, Coprococcus_catus, Eubacterium_ventriosum, Intestinibacter_bartlettii, Enterococcus_faecium, Gordonibacter_pamelaeae, Clostridium_citroniae, Dialister_succinatiphilus, Sutterella_wadsworthensis, Paraprevotella_clara, Lactobacillus_ruminis, Lactobacillus_crispatus, Parasutterella_excrementihominis, Clostridium_leptum, Limosilactobacillus_reuteri, Bacteroides_dorei, Bifidobacterium_breve, Bacteroides_caccae, Bacteroides_coprocola, Ruminococcus_lactaris, Parabacteroides_merdae, Veillonella_parvula, Bacteroides_finegoldii, Escherichia_coli, Bacteroides_coprophilus, Odoribacter_laneus, Megamonas_hypermegale, Desulfovibrio_piger, Paraprevotella_xylaniphila, Collinsella_intestinalis, Prevotella_copri, Bacteroides_vulgatus, Bacteroides_stercoris, Bacteroides_uniformis, Prevotella_stercorea, Roseburia_inulinivorans, Roseburia_intestinalis, Bacteroides_thetaiotaomicron, Bacteroides_intestinalis, Roseburia_hominis, Alistipes_shahii, Bifidobacterium_pseudocatenulatum, Bacteroides_xylanisolvens, Bacteroides_cellulosilyticus, Oxalobacter_formigenes, Blautia_producta, Bilophila_wadsworthia, Ruminococcus_bromii, Bifidobacterium_Bifidum, Coprococcus_eutactus, Clostridium_hathewayi, Tyzzerella_nexilis, Anaerotruncus_colihominis, Holdemania_filiformis, Alistipes_indistinctus, Collinsella_aerofaciens, Faecalibacterium_prausnitzii, Alistipes_putredinis, Collinsella_tanakaei, Bacteroides_eggerthii, Acidaminococcus_intestini, Turicibacter_sanguinis, Bacteroides_ovatus, Megamonas_funiformis, Methanobrevibacter_smithii, Butyrivibrio_crossotus, Bifidobacterium_angulatum, Parabacteroides_johnsonii, Eubacterium_siraeum, Citrobacter_freundii, Akkermansia_muciniphila, Bacteroides_fragilis, Odoribacter_splanchnicus, Bacteroides_plebeius, Bacteroides_clarus, Clostridium_symbiosum, Clostridium_asparagiforme, Catenibacterium_mitsuokai, Bacteroides_distasonis, Fusobacterium_mortiferum, Enterobacter_cloacae, Megasphaera_elsdenii, Blautia_hydrogenotrophica, Bacteroides_fluxus, Bifidobacterium_dentium, Clostridium_bolteae , Clostridium_difficile, Lactobacillus_delbrueckii, Lactobacillus_johnsonii, Lactobacillus_fermentum, Lactobacillus_gasseri, Fusobacterium_varium, Enterococcus_faecalis, Streptococcus_infantarius, Streptococcus_anginosus, Lactobacillus_salivarius, Ruminococcus_gnavus, Fusobacterium_ulcerans, Clostridium_orbiscindens, Bifidobacterium_animalis, Mitsuokella_multacida, Blautia_hansenii, Lactobacillus_casei, Anaerostipes_caccae, Clostridium_scindens, Subdoligranulum_variabile, Clostridium_spiroforme, Collinsella_stercoris, Lactobacillus_acidophilus, Lacticaseibacillus_rhamnosus, Pediococcus_acidilactici, Pseudoflavonifractor_capillosus, Pseudomonas_sp., Sutterella_parvirubra, Eggerthella._sp., Clostridium_sp., Subdoligranulum_sp., Clostridium_sporogenes, Succinatimonas_hippei, Butyrivibrio_fibrisolvens, Parabacteroides_sp., Lactobacillus_paracasei, Finegoldia_magna, Veillonella_sp., Pseudomonas_aeruginosa, Ruminococcus_sp., Fusobacterium_sp., Bacteroides_sp., Bacillus_sp., Enterococcus_sp., Escherichia_sp., Anaerococcus_hydrogenalis, Proteus_penneri, Providencia_rettgeri, Edwardsiella_tarda, Acinetobacter_junii, Neisseria_macacae, Acinetobacter_baumannii, Bacillus_cereus, Campylobacter_hominis, Enterococcus_saccharolyticus, Campylobacter_jejuni, Campylobacter_curvus, Campylobacter_concisus, Helicobacter_pylori, Lactobacillus_helveticus, Listeria_monocytogenes, Porphyromonas_gingivalis, Yersinia_enterocolitica, Yersinia_pseudotuberculosis, Helicobacter_bilis, Providencia_stuartii, Listeria_grayi, Lactobacillus_buchneri, Lactobacillus_brevis, Lactobacillus_antri, Leuconostoc_mesenteroides, Streptococcus_equinus, Ruminococcaceae_bacterium, Laribacter_hongkongensis, Yersinia_bercovieri, Vibrio_mimicus, Vibrio_furnissii, Escherichia_albertii, Hafnia_alvei, Grimontia_hollisae, Corynebacterium_ammoniagenes, Clostridium_hiranonis, Lachnospiraceae_bacterium, Bifidobacterium_gallicum, Lactobacillus_amylolyticus, Lactobacillus_ultunensis, Marvinbryantia_formatexigens, Bacteroides_pectinophilus, Erysipelotrichaceae_bacterium, Prevotella_salivae, Clostridium_hylemonae, Citrobacter_koseri, Citrobacter_youngae, Enterobacteriaceae_bacterium, Lactobacillus_coryniformis, Arcobacter_butzleri, Burkholderiales_bacterium, Enterobacter_cancerogenus, Providencia_alcalifaciens, Victivallis_vadensis, Parvimonas_micra, Anaerofustis_stercorihominis, Dysgonomonas_mossii, Lactobacillus_fructivorans, Campylobacter_upsaliensis, Fusobacterium_necrophorum, Fusobacterium_gonidiaformans, NA, NA, Paenibacillus_sp., NA, NA, Vibrio_cholerae, NA, NA, Providencia_rustigianii, NA, NA, NA, NA, Bifidobacterium_sp., NA, Campylobacter_coli, NA, NA, NA, Clostridium_methylpentosum, Coprobacillus_sp., NA, Desulfovibrio_sp., NA, Helicobacter_canadensis, Helicobacter_cinaedi, Helicobacter_pullorum, Helicobacter_winghamensis, NA, Methanosphaera_stadtmanae, NA, NA, Streptococcus_sp., NA, Eubacterium_sp., NA, Weissella_paramesenteroides, NA, Bartonella_quintana, NA